+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog], and this project adheres to [Semantic Versioning].
 
 ## [Unreleased]
+### Changed
+- Update default app URL to `https://unforgettable.app/sdk` across all platforms
+
+### Removed
+- **BREAKING**: Remove `helperDataUrl` from all SDK packages (core, android, ios, react)
 
 ## [0.8.0] - 2025-11-28
 ### Added
@@ -75,7 +80,8 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/rarimo/unforgettable-sdk/compare/0.8.0...HEAD
+[Unreleased]: https://github.com/rarimo/unforgettable-sdk/compare/0.9.0...HEAD
+[0.9.0]: https://github.com/rarimo/unforgettable-sdk/compare/0.8.0...0.9.0
 [0.8.0]: https://github.com/rarimo/unforgettable-sdk/compare/0.7.0...0.8.0
 [0.7.0]: https://github.com/rarimo/unforgettable-sdk/compare/0.6.0...0.7.0
 [0.6.0]: https://github.com/rarimo/unforgettable-sdk/compare/0.5.1...0.6.0